Revert "fix(query): avoid exchange coordinator reentrant deadlock"
This reverts commit bd0bbdc7da3eeab47c97a046c9480486cf1d48b3.
